Moved the MyActivity.java file out of the loop completely. When loading the app now we get the first RSS feed that cnn top stories has available. Clicking the TextView works as well. Loads in the default browser.
TODO: I need to change HandleXML.java to get all the items on cnn. Set item = true when we find a start element of item. We then get the next title, link, description and push that to infos[count]. When we find the end tag of item we (in this order)  rssInfos.add(count,info)...count++.....item = false. This way we only get the title etc when we are known to be inside the item tag. When we reach the end tag we should have all the stuff we need so add it to the arrayList.